# frontend/packages/types/openapi.json
{
  "components": {
    "schemas": {
      "AIConfigRead": {
        "properties": {
          "chat": {
            "$ref": "#/components/schemas/ProviderRead"
          },
          "embedding": {
            "$ref": "#/components/schemas/ProviderRead"
          },
          "embedding_dimension": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Dimension"
          },
          "embedding_locked": {
            "title": "Embedding Locked",
            "type": "boolean"
          }
        },
        "required": [
          "chat",
          "embedding",
          "embedding_dimension",
          "embedding_locked"
        ],
        "title": "AIConfigRead",
        "type": "object"
      },
      "AIConfigTest": {
        "additionalProperties": false,
        "description": "An in-flight configuration, not necessarily the saved one.\n\nCredentials are optional for the same reason they are optional on a save: a test should\nexercise *what would be saved*, and for an omitted credential that is the one already\nstored. Requiring them here would mean re-typing a key that cannot be read back merely to\ncorrect a model name. The service resolves the fallback and only accepts it when the\nprovider still matches \u2014 a key stored for Azure proves nothing about Bedrock.",
        "properties": {
          "chat": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatTarget"
              },
              {
                "type": "null"
              }
            ]
          },
          "embedding": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmbeddingTarget"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "AIConfigTest",
        "type": "object"
      },
      "AIConfigTestResult": {
        "properties": {
          "embedding_dimension": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Dimension"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "failed": {
            "anyOf": [
              {
                "enum": [
                  "chat",
                  "embedding"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failed"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "AIConfigTestResult",
        "type": "object"
      },
      "AIConfigUpdate": {
        "additionalProperties": false,
        "description": "A whole configuration, replaced in one call.\n\nCredentials may be omitted to keep the ones already stored, which is what lets a model or\nan endpoint be corrected without re-typing a key that was never readable back.",
        "properties": {
          "chat": {
            "$ref": "#/components/schemas/ChatTarget"
          },
          "embedding": {
            "$ref": "#/components/schemas/EmbeddingTarget"
          }
        },
        "required": [
          "chat",
          "embedding"
        ],
        "title": "AIConfigUpdate",
        "type": "object"
      },
      "AcceptInvitationRequest": {
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "password": {
            "description": "At least 10 characters",
            "maxLength": 256,
            "title": "Password",
            "type": "string"
          },
          "token": {
            "maxLength": 256,
            "minLength": 16,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "password"
        ],
        "title": "AcceptInvitationRequest",
        "type": "object"
      },
      "AcceptInvitationResponse": {
        "properties": {
          "organization": {
            "$ref": "#/components/schemas/OrganizationRead"
          },
          "tokens": {
            "$ref": "#/components/schemas/TokenPair"
          },
          "user": {
            "$ref": "#/components/schemas/UserRead"
          }
        },
        "required": [
          "organization",
          "user",
          "tokens"
        ],
        "title": "AcceptInvitationResponse",
        "type": "object"
      },
...
      "ChatConnection": {
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Version"
          },
          "base_url": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "endpoint": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endpoint"
          },
          "region": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "think": {
            "default": true,
            "title": "Think",
            "type": "boolean"
          }
        },
        "title": "ChatConnection",
        "type": "object"
      },
      "ChatProviderName": {
        "enum": [
          "azure",
          "bedrock",
          "anthropic",
          "ollama"
        ],
        "title": "ChatProviderName",
        "type": "string"
      },
      "ChatTarget": {
        "additionalProperties": false,
        "description": "One half of a configuration: what to call and how to reach it.",
        "properties": {
          "connection": {
            "$ref": "#/components/schemas/ChatConnection"
          },
          "credentials": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProviderCredentials"
              },
              {
                "type": "null"
              }
            ]
          },
          "model": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "$ref": "#/components/schemas/ChatProviderName"
          }
        },
        "required": [
          "provider",
          "model"
        ],
        "title": "ChatTarget",
        "type": "object"
      },
...
      "EmbeddingProviderName": {
        "description": "Anthropic is absent on purpose \u2014 it publishes no embeddings API.\n\nIts omission here is what the API's 422, the database check constraint and the factory's\ndispatch all derive from, so there is one place to change if that ever stops being true.",
        "enum": [
          "azure",
          "bedrock",
          "ollama"
        ],
        "title": "EmbeddingProviderName",
        "type": "string"
      },
      "EmbeddingTarget": {
        "additionalProperties": false,
        "properties": {
          "connection": {
            "$ref": "#/components/schemas/ProviderConnection"
          },
          "credentials": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProviderCredentials"
              },
              {
                "type": "null"
              }
            ]
          },
          "model": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "$ref": "#/components/schemas/EmbeddingProviderName"
          }
        },
        "required": [
          "provider",
          "model"
        ],
        "title": "EmbeddingTarget",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvitationPreview": {
        "description": "What the sign-up page may show before the invitee has proved anything.\n\nThe organisation name and the role are enough to make the page meaningful; the inviter's\nidentity and the rest of the team are not disclosed to someone holding only a link.",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "organization_name": {
            "title": "Organization Name",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "required": [
          "organization_name",
          "email",
          "role",
          "expires_at"
        ],
        "title": "InvitationPreview",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "OrganizationRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "plan": {
            "$ref": "#/components/schemas/Plan"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "plan",
          "created_at"
        ],
        "title": "OrganizationRead",
        "type": "object"
      },
...
      "Plan": {
        "enum": [
          "free",
          "pro",
          "enterprise"
        ],
        "title": "Plan",
        "type": "string"
      },
...
      "ProviderCredentials": {
        "additionalProperties": false,
        "description": "Write-only, in the same sense the chatbot's secret key is: it goes in and never comes\nback out. Which members matter depends on the provider \u2014 see `registry.py`.",
        "properties": {
          "access_key_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Key Id"
          },
          "api_key": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "secret_access_key": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret Access Key"
          }
        },
        "title": "ProviderCredentials",
        "type": "object"
      },
      "ProviderRead": {
        "description": "What the dashboard is told about one half. No credential ever appears here.",
        "properties": {
          "connection": {
            "additionalProperties": true,
            "title": "Connection",
            "type": "object"
          },
          "credentials_set": {
            "title": "Credentials Set",
            "type": "boolean"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "ready": {
            "title": "Ready",
            "type": "boolean"
          }
        },
        "required": [
          "provider",
          "model",
          "connection",
          "credentials_set",
          "ready"
        ],
        "title": "ProviderRead",
        "type": "object"
      },
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
      "SignupRequest": {
        "description": "Creates the organisation and its first owner in one step.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "organization_name": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Organization Name",
            "type": "string"
          },
          "password": {
            "description": "At least 10 characters",
            "maxLength": 256,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "organization_name",
          "email",
          "password"
        ],
        "title": "SignupRequest",
        "type": "object"
      },
      "SignupResponse": {
        "properties": {
          "organization": {
            "$ref": "#/components/schemas/OrganizationRead"
          },
          "tokens": {
            "$ref": "#/components/schemas/TokenPair"
          },
          "user": {
            "$ref": "#/components/schemas/UserRead"
          }
        },
        "required": [
          "organization",
          "user",
          "tokens"
        ],
        "title": "SignupResponse",
        "type": "object"
      },
...
      "TokenPair": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "expires_in"
        ],
        "title": "TokenPair",
        "type": "object"
      },
      "UserRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "org_id": {
            "format": "uuid",
            "title": "Org Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "required": [
          "id",
          "org_id",
          "email",
          "full_name",
          "role",
          "is_active",
          "created_at"
        ],
        "title": "UserRead",
        "type": "object"
      },
      "UserRole": {
        "enum": [
          "owner",
          "admin",
          "member"
        ],
        "title": "UserRole",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Dashboard access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "RAG Chatbot Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/invitations/accept": {
      "post": {
        "operationId": "accept_invitation_api_v1_auth_invitations_accept_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptInvitationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Invitation",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/invitations/preview": {
      "get": {
        "description": "Lets the acceptance page name the organisation before the invitee sets a password.",
        "operationId": "preview_invitation_api_v1_auth_invitations_preview_get",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 16,
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationPreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Invitation",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Retires the refresh token so the session cannot be renewed.\n\nUnauthenticated on purpose: possession of the refresh token is the only thing being\nacted on, and a client whose access token has already expired must still be able to sign\nout cleanly.",
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Me",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/signup": {
      "post": {
        "operationId": "signup_api_v1_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Signup",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/chatbots/{chatbot_id}/ai-config": {
      "get": {
        "description": "Providers, models and connection detail. Credentials are write-only and never returned,\nthe same posture as the chatbot's secret key.",
        "operationId": "get_ai_config_api_v1_chatbots__chatbot_id__ai_config_get",
        "parameters": [
          {
            "in": "path",
            "name": "chatbot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chatbot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIConfigRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Ai Config",
        "tags": [
          "ai-config"
        ]
      },
      "put": {
        "description": "Replace the whole configuration.\n\nOmit a `credentials` object to keep the key already stored. Changing the embedding\nprovider or model is refused with a 409 once the chatbot has chunks: existing vectors\ncannot be compared against a different model's.",
        "operationId": "put_ai_config_api_v1_chatbots__chatbot_id__ai_config_put",
        "parameters": [
          {
            "in": "path",
            "name": "chatbot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chatbot Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIConfigRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Put Ai Config",
        "tags": [
          "ai-config"
        ]
      }
    },
    "/api/v1/chatbots/{chatbot_id}/ai-config/test": {
      "post": {
        "description": "Call the providers with values that have not been saved yet.\n\nAlways 200 \u2014 the verdict is in the body, because a provider rejecting a tenant's key is a\nresult this endpoint delivered successfully, not a failure of this endpoint.",
        "operationId": "test_ai_config_api_v1_chatbots__chatbot_id__ai_config_test_post",
        "parameters": [
          {
            "in": "path",
            "name": "chatbot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chatbot Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIConfigTest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIConfigTestResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Test Ai Config",
        "tags": [
          "ai-config"
        ]
      }
    },
...
  }
}